Refactor code

+++ src/viewport.h
@@ -1,16 +1,14 @@
 #pragma once
 #include "plugin.h"
 
-enum class eViewportState
-{
+enum class eViewportState {
     Edit,
     View,
     Browser,
 };
 
-class Viewport
-{
-private:
+class Viewport {
+  private:
     static inline bool m_bInitialized; // Is the viewport init done
     static inline ImVec2 m_fSize; // Size of the viewport
     static inline CVector m_fMousePos; // Mouse position in the viewport
@@ -33,27 +31,26 @@ private:
     // TODO: Should be in ObjManager class?
     static void ProcessInputs();
 
-public:
+  public:
     static inline eViewportState m_eState = eViewportState::Edit;
     static inline bool m_bHovered; // Is mouse hovering viewport
     static inline CEntity *m_HoveredEntity; // Currently hovered entity
-    static inline int m_nMoveSpeed = 1; // movement speed multiplier controls
+    static inline float m_nMoveSpeed = 1.0f; // movement speed multiplier controls
     static inline float m_fFOV = 70.0f;
     static inline CVector m_fWorldPos; // cursor world position
 
-    struct Browser
-    {
-    private:
+    struct Browser {
+      private:
         static inline size_t m_nSelectedID = NULL; // Selected model id
         static inline CVector m_fRot; // rotation of the rendering object
 
         // Loads model
         // TODO: Need some fail checks
         static void LoadModel(size_t model, RpClump *&pClump, RpAtomic *&pAtomic, RwFrame *&pFrame);
-        
+
         // Renders the model in viewport
         static void RenderModel();
-    public:
+      public:
         static inline bool m_bShowNextFrame; // Should the browser be shown next frame
         static inline bool m_bShown; // Is the browser being shown
         static inline bool m_bAutoRot; // Auto rotate object in browser
